Extract game logic to manager
