+++ mj-calc/src/modules/Calculator.tsx
@@ -1,25 +1,19 @@
 import { useCallback, useContext, useMemo, useState } from "react";
-import { includes, maxBy, omit, pick } from "lodash";
+import { includes, maxBy } from "lodash";
 
 import PlayerTable from "./components/PlayerTable";
 import { PlayerList, Player } from "./types/Player";
 import { getWind, WindNumber } from "./util/Wind";
 import { incrementRound } from "./util/GameStatus";
 import { GameStatus } from "./types/GameStatus";
-import {
-  applyScoreChange,
-  getDealer,
-  getDeltaWithoutWinner,
-  getDeltaWithWinner,
-} from "./util/Score";
+import { getDealer } from "./util/Score";
 import { HonbaStick, RichiiStick } from "./Icons";
-import { Record } from "./types/Record";
 import RoundHistory from "./components/RoundHistory";
 import { GameEntrySelector } from "./components/GameEntrySelector";
 import GameContext from "./context/GameContext";
 import FinalPoints from "./components/FinalPoints";
 import { PointsPlot } from "./components/PointsPlot";
-import { DEFAULT_WIN_INFO, StickIconSize } from "./util/Constants";
+import { StickIconSize } from "./util/Constants";
 import { useGameManager } from "./hooks/useGameManager";
 import { useToggle } from "./hooks/useToggle";
 import { ResultInputContext } from "./context/ResultInputContext";
@@ -43,37 +37,14 @@ export function Calculator({
   onNextGame: (names: string[]) => void;
 }) {
   // Game state objects
-  const {
-    gameStatus,
-    setGameStatus,
-    players,
-    setPlayers,
-    records: gameRecord,
-    setRecords: setGameRecord,
-  } = useContext(GameContext);
+  const { gameStatus, players, records: gameRecord } = useContext(GameContext);
   const n = gameStatus.numPlayers;
 
   // New result input
   const { winInfo, setWinInfo, tenpai, setTenpai, endingType, setEndingType } =
     useContext(ResultInputContext);
 
-  const { togglePlayerRichii } = useGameManager();
-
-  const pushRecord = useCallback(
-    (record: Record) => {
-      setGameRecord((gameRecord) => [...gameRecord, record]);
-    },
-    [setGameRecord]
-  );
-
-  const resetWinState = useCallback(
-    (gameStatus: GameStatus) => {
-      setWinInfo(DEFAULT_WIN_INFO(gameStatus));
-      setTenpai(Array(n).fill(false));
-      setEndingType("Win");
-    },
-    [n, setEndingType, setTenpai, setWinInfo]
-  );
+  const { togglePlayerRichii, rewind, saveEntry } = useGameManager();
 
   // Display
   const [displayDelta, setDisplayDelta] = useState(-1);
@@ -90,98 +61,6 @@ export function Calculator({
       }
     });
   }
-
-  // Rewind
-  const rewind = useCallback(() => {
-    // Pop last record
-    setGameRecord((gameRecord) => {
-      const newRecords = [...gameRecord];
-      const lastRecord = newRecords.pop();
-      if (lastRecord === undefined) {
-        // Rewind triggered when no game has been played yet
-        return newRecords;
-      }
-
-      // Replace with last status
-      setGameStatus((gameStatus) => {
-        // Update players' score based on ending and richii
-        setPlayers((players) =>
-          applyScoreChange(
-            players,
-            lastRecord.deltas.map(
-              (delta, index) => (gameStatus.richii[index] ? 1000 : 0) - delta
-            )
-          )
-        );
-
-        return {
-          numPlayers: gameStatus.numPlayers,
-          ...pick(lastRecord, "wind", "round", "honba", "richiiStick"),
-          richii: [...lastRecord.richii],
-        };
-      });
-
-      setEndingType(lastRecord.type);
-      if (lastRecord.type === "Win") {
-        setWinInfo(lastRecord.info);
-      } else {
-        setTenpai(lastRecord.info);
-      }
-      return newRecords;
-    });
-  }, [
-    setGameRecord,
-    setGameStatus,
-    setEndingType,
-    setPlayers,
-    setWinInfo,
-    setTenpai,
-  ]);
-
-  const saveEntry = useCallback(() => {
-    setGameStatus((gameStatus) => {
-      let deltas: number[];
-      if (endingType === "Win") {
-        deltas = getDeltaWithWinner(winInfo, gameStatus);
-      } else {
-        deltas = getDeltaWithoutWinner(tenpai);
-      }
-      setPlayers((players) => applyScoreChange(players, deltas));
-      const record: Omit<Record, "info" | "type"> = {
-        deltas,
-        ...omit(gameStatus, "numPlayers"),
-      };
-      if (endingType === "Win") {
-        pushRecord({
-          ...record,
-          type: endingType,
-          info: winInfo,
-        });
-      } else {
-        pushRecord({
-          ...record,
-          type: endingType,
-          info: tenpai,
-        });
-      }
-
-      resetWinState(gameStatus);
-
-      return nextGameStatus(
-        endingType === "Win" ? winInfo.map((record) => record.winner) : null,
-        tenpai[getDealer(gameStatus)],
-        gameStatus
-      );
-    });
-  }, [
-    endingType,
-    pushRecord,
-    resetWinState,
-    setGameStatus,
-    setPlayers,
-    tenpai,
-    winInfo,
-  ]);
 
   const PlayerInfoCell = useCallback(
     (player: Player) => {
