+++ mj-calc/src/modules/context/ResultInputContext.ts
@@ -1,5 +1,6 @@
 import React from "react";
 import { Dispatch, SetStateAction } from "react";
+import { GameStatus } from "../types/GameStatus";
 import { WinRecord } from "../types/Record";
 
 interface ResultInputContextType {
@@ -9,6 +10,7 @@ interface ResultInputContextType {
   setEndingType: Dispatch<SetStateAction<"Win" | "Draw">>;
   tenpai: boolean[];
   setTenpai: Dispatch<SetStateAction<boolean[]>>;
+  resetWinState: (GameStatus: GameStatus) => void;
 }
 
 export const ResultInputContext = React.createContext<ResultInputContextType>(
