+++ mj-calc/src/modules/provider/ResultInputContextProvider.tsx
@@ -1,6 +1,7 @@
-import { useContext, useMemo, useState } from "react";
+import { useContext, useMemo, useState, useCallback } from "react";
 import GameContext from "../context/GameContext";
 import { ResultInputContext } from "../context/ResultInputContext";
+import { GameStatus } from "../types/GameStatus";
 import { WinRecord } from "../types/Record";
 import { DEFAULT_WIN_INFO } from "../util/Constants";
 
@@ -11,7 +12,7 @@ export function ResultInputContextProvider({
   children,
 }: ResultInputContextProps) {
   const { gameStatus } = useContext(GameContext);
-  
+
   const [winInfo, setWinInfo] = useState<WinRecord[]>(
     DEFAULT_WIN_INFO(gameStatus)
   );
@@ -19,6 +20,15 @@ export function ResultInputContextProvider({
     Array(gameStatus.numPlayers).fill(false)
   );
   const [endingType, setEndingType] = useState<"Win" | "Draw">("Win");
+
+  const resetWinState = useCallback(
+    (gameStatus: GameStatus) => {
+      setWinInfo(DEFAULT_WIN_INFO(gameStatus));
+      setTenpai(Array(gameStatus.numPlayers).fill(false));
+      setEndingType("Win");
+    },
+    [setEndingType, setTenpai, setWinInfo]
+  );
 
   const contextValue = useMemo(
     () => ({
@@ -28,8 +38,9 @@ export function ResultInputContextProvider({
       setEndingType,
       tenpai,
       setTenpai,
+      resetWinState,
     }),
-    [endingType, tenpai, winInfo]
+    [endingType, resetWinState, tenpai, winInfo]
   );
 
   return (
